// react-next-jest/src/app/interfaces/infoEmpresa.ts
export default interface InfoEmpresaInterface {
    CIF: string;
    dirección: string;
    localidad: string;
    teléfono: string;
    email: string;
    idiomaRelacional: string;
    sector: string;
    corporacionMondragon: string;
    numeroTrabajadores: string;
    cursoAcademico: string;
    fechaPrimeraVisitaCaptacion: string;
  }